Altered RetryTime time for xBridge.
I found the 65 second retry time was too long for xBridge at times.  Reducing it to 25 seconds has given a good boost to performance.  I believe that Android often gets too busy to allow the retry timer to be reliable, and sometimes xDrip does not re-establish the connection to a newly woken xBridge in time to get the data packet.  This change has proven this thought to be somewhat correct.

I decided not to make it shorter than this to keep the logs from filling up with messages about it.