Enabled high precision Delta String with units.
NOTE: DO NOT MERGE THIS UNTIL A NEW WATCH FACE IS PUSHED!!!!

Merging this into xDrip Beta or Release will break the watch face for users.  The current watch face is expecting only a numerical string with 2 decimal places, and  no units.
I have created a "dumber" watch face that will need to be pushed to the Pebble Store before this is merged and put into general circulation.  This watch face simply accepts the full unitized delta string with units and displays it as is.  No fancy conversion to Integer and determining if it has a decimal point in it to then add the units.  A much better way to do things.  Sorry it took me so long to get to this.