In xbridge battery level = 0, then don't show the red status message in home screen saying 'waiting for packet'
- I'm using an xdrip with xbridge software installed, but without the resistor wiring
As a consequence, batterylevel is always 0, but I always see on the Home screen in red "waiting for packet"
So I removed that : if batterylevel = 0 then don't show that message at all. For an xBridge that has the wiring, I assume battery level should never be 0 or it's dead.

Moreover, it looks a strange message "waiting for package ?"